atualização pedido

+++ src/main/java/com/example/demo/repository/UsuarioRepository.java
@@ -11,8 +11,8 @@ public interface UsuarioRepository extends CrudRepository<Usuario,Integer> {
 
     Optional<Usuario> findByNome(String nome);
 
-    Optional<Usuario>findById(Long id);
+    Optional<Usuario>findById(Integer id);
 
-    boolean existsUsuarioById(Long id);
+    boolean existsUsuarioById(Integer id);
 
 }
